+++ spec/spec_helper.rb
@@ -29,7 +29,6 @@ end
 
 PIP_VERSION = get_requirement_version('pip')
 SETUPTOOLS_VERSION = get_requirement_version('setuptools')
-WHEEL_VERSION = get_requirement_version('wheel')
 PIPENV_VERSION = get_requirement_version('pipenv')
 POETRY_VERSION = get_requirement_version('poetry')
 UV_VERSION = get_requirement_version('uv')